# blog-gin/docs/swagger.yaml
definitions:
...
  dto.CategoryBackDTO:
    properties:
      article_count:
        type: integer
      category_name:
        description: 分类名
        type: string
      created_at:
        description: 创建时间
        type: integer
      id:
        type: integer
      updated_at:
        description: 更新时间
        type: integer
    type: object
...
  dto.CategoryNewReq:
    properties:
      category_name:
        description: 分类名
        type: string
      id:
        type: integer
    type: object
  dto.CategoryQuery:
    properties:
      category_name:
        description: 分类名
        type: string
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      sorts:
        description: 排序
        items:
          type: string
        type: array
    type: object
...
  dto.PhotoBackDTO:
    properties:
      album_id:
        description: 相册id
        type: integer
      created_at:
        description: 创建时间
        type: integer
      id:
        description: 主键
        type: integer
      is_delete:
        description: 是否删除
        type: integer
      photo_desc:
        description: 照片描述
        type: string
      photo_name:
        description: 照片名
        type: string
      photo_src:
        description: 照片地址
        type: string
      updated_at:
        description: 更新时间
        type: integer
    type: object
  dto.PhotoNewReq:
    properties:
      album_id:
        description: 相册id
        type: integer
      id:
        description: 主键
        type: integer
      is_delete:
        description: 是否删除
        type: integer
      photo_desc:
        description: 照片描述
        type: string
      photo_name:
        description: 照片名
        type: string
      photo_src:
        description: 照片地址
        type: string
    type: object
  dto.PhotoQuery:
    properties:
      album_id:
        description: 相册id
        type: integer
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      sorts:
        description: 排序
        items:
          type: string
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.BatchResp:
    properties:
      success_count:
        type: integer
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdReq:
    properties:
      id:
        type: integer
    type: object
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.PageResp:
    properties:
      list: { }
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: { }
      message:
        type: string
      trace_id:
        type: string
    type: object
host: localhost:9090
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support/
  description: APP接口鉴权方式：token采用自加密token的方式，自加密token规则：sha256(${x-timestamp}${base_secret})
    生成的加密token，x-timestamp 时间戳需要保持10分钟内
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger Example API
  version: 1.0.0
paths:
...
  /admin_api/v1/album/batch_delete_photo:
    delete:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdsReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.BatchResp'
                type: object
      summary: '"批量删除照片"'
      tags:
        - Photo
...
  /admin_api/v1/category/add_category:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.CategoryNewReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.CategoryBackDTO'
                type: object
      summary: '"创建文章分类"'
      tags:
        - Category
  /admin_api/v1/category/batch_delete_category:
    delete:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdsReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.BatchResp'
                type: object
      summary: '"批量删除文章分类"'
      tags:
        - Category
  /admin_api/v1/category/delete_category:
    delete:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.BatchResp'
                type: object
      summary: '"删除文章分类"'
      tags:
        - Category
  /admin_api/v1/category/find_category_list:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.CategoryQuery'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.PageResp'
                type: object
      summary: '"分页获取文章分类列表"'
      tags:
        - Category
  /admin_api/v1/category/update_category:
    put:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.CategoryNewReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.CategoryBackDTO'
                type: object
      summary: '"更新文章分类"'
      tags:
        - Category
...
  /admin_api/v1/photo/add_photo:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.PhotoNewReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.PhotoBackDTO'
                type: object
      summary: '"创建照片"'
      tags:
        - Photo
  /admin_api/v1/photo/delete_photo:
    delete:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.IdReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.BatchResp'
                type: object
      summary: '"删除照片"'
      tags:
        - Photo
  /admin_api/v1/photo/find_photo_list:
    post:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.PhotoQuery'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/github_com_ve-weiyi_ve-blog-golang_blog-gin_service_admin_dto.PageResp'
                type: object
      summary: '"分页获取照片列表"'
      tags:
        - Photo
  /admin_api/v1/photo/update_photo:
    put:
      consumes:
        - application/json
      parameters:
        - description: 请求参数
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/dto.PhotoNewReq'
      produces:
        - application/json
      responses:
        "200":
          description: 返回信息
          schema:
            allOf:
              - $ref: '#/definitions/response.Body'
              - properties:
                  data:
                    $ref: '#/definitions/dto.PhotoBackDTO'
                type: object
      summary: '"更新照片"'
      tags:
        - Photo
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"